# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Platform Engineering Control Plane API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export interface User {
  id: number;
  githubId: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  email: string | null;
  role: UserRole;
  createdAt: string;
}

export interface Team {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  createdBy: number;
  createdAt: string;
  updatedAt: string;
  memberCount: number;
  namespaceCount: number;
}

export interface CreateTeamBody {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateTeamBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export type TeamMemberRole =
  (typeof TeamMemberRole)[keyof typeof TeamMemberRole];

export const TeamMemberRole = {
  owner: "owner",
  member: "member",
} as const;

export interface TeamMember {
  id: number;
  teamId: number;
  userId: number;
  role: TeamMemberRole;
  joinedAt: string;
  user: User;
}

export type AddTeamMemberBodyRole =
  (typeof AddTeamMemberBodyRole)[keyof typeof AddTeamMemberBodyRole];

export const AddTeamMemberBodyRole = {
  owner: "owner",
  member: "member",
} as const;

export interface AddTeamMemberBody {
  userId: number;
  role?: AddTeamMemberBodyRole;
}

export type NamespaceEnvironment =
  (typeof NamespaceEnvironment)[keyof typeof NamespaceEnvironment];

export const NamespaceEnvironment = {
  dev: "dev",
  prod: "prod",
} as const;

export type NamespaceStatus =
  (typeof NamespaceStatus)[keyof typeof NamespaceStatus];

export const NamespaceStatus = {
  active: "active",
  provisioning: "provisioning",
  terminating: "terminating",
} as const;

export type NamespaceResourceQuota = {
  cpuLimit: string;
  memoryLimit: string;
  podLimit: number;
};

export interface Namespace {
  id: number;
  teamId: number;
  teamName: string;
  teamSlug: string;
  environment: NamespaceEnvironment;
  k8sNamespace: string;
  status: NamespaceStatus;
  createdAt: string;
  resourceQuota: NamespaceResourceQuota;
}

export type CreateNamespaceBodyEnvironment =
  (typeof CreateNamespaceBodyEnvironment)[keyof typeof CreateNamespaceBodyEnvironment];

export const CreateNamespaceBodyEnvironment = {
  dev: "dev",
  prod: "prod",
} as const;

export interface CreateNamespaceBody {
  teamId: number;
  environment: CreateNamespaceBodyEnvironment;
}

export type StarterKitCategory =
  (typeof StarterKitCategory)[keyof typeof StarterKitCategory];

export const StarterKitCategory = {
  web: "web",
  api: "api",
  worker: "worker",
  ml: "ml",
  data: "data",
} as const;

export interface StarterKit {
  id: number;
  name: string;
  description: string;
  category: StarterKitCategory;
  language: string;
  framework: string;
  tags: string[];
  dockerfileIncluded: boolean;
  cicdIncluded: boolean;
  complianceReady: boolean;
  createdAt: string;
}

export type DeploymentEnvironment =
  (typeof DeploymentEnvironment)[keyof typeof DeploymentEnvironment];

export const DeploymentEnvironment = {
  dev: "dev",
  prod: "prod",
} as const;

export type DeploymentStatus =
  (typeof DeploymentStatus)[keyof typeof DeploymentStatus];

export const DeploymentStatus = {
  pending: "pending",
  running: "running",
  blocked: "blocked",
  failed: "failed",
  succeeded: "succeeded",
} as const;

export type DeploymentComplianceStatus =
  (typeof DeploymentComplianceStatus)[keyof typeof DeploymentComplianceStatus];

export const DeploymentComplianceStatus = {
  pending: "pending",
  passed: "passed",
  failed: "failed",
} as const;

export interface Deployment {
  id: number;
  teamId: number;
  teamName: string;
  namespaceId: number;
  namespaceName: string;
  environment: DeploymentEnvironment;
  serviceName: string;
  imageTag: string;
  status: DeploymentStatus;
  complianceStatus: DeploymentComplianceStatus;
  /** @nullable */
  testCoverage: number | null;
  triggeredBy: number;
  triggeredByUsername: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeploymentBody {
  namespaceId: number;
  serviceName: string;
  imageTag: string;
  /** @nullable */
  testCoverage?: number | null;
}

export type EvidenceRecordEnvironment =
  (typeof EvidenceRecordEnvironment)[keyof typeof EvidenceRecordEnvironment];

export const EvidenceRecordEnvironment = {
  dev: "dev",
  prod: "prod",
} as const;

export type EvidenceRecordCheckType =
  (typeof EvidenceRecordCheckType)[keyof typeof EvidenceRecordCheckType];

export const EvidenceRecordCheckType = {
  test_coverage: "test_coverage",
  security_scan: "security_scan",
  policy_gate: "policy_gate",
  rbac_check: "rbac_check",
} as const;

export type EvidenceRecordStatus =
  (typeof EvidenceRecordStatus)[keyof typeof EvidenceRecordStatus];

export const EvidenceRecordStatus = {
  passed: "passed",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface EvidenceRecord {
  id: number;
  deploymentId: number;
  teamId: number;
  teamName: string;
  environment: EvidenceRecordEnvironment;
  serviceName: string;
  imageTag: string;
  checkType: EvidenceRecordCheckType;
  status: EvidenceRecordStatus;
  details: string;
  /** @nullable */
  threshold: number | null;
  /** @nullable */
  actual: number | null;
  createdAt: string;
}

export type OperatorType = (typeof OperatorType)[keyof typeof OperatorType];

export const OperatorType = {
  s3: "s3",
  rds: "rds",
  salesforce: "salesforce",
  kafka: "kafka",
  redis: "redis",
  vault: "vault",
} as const;

export type OperatorStatus =
  (typeof OperatorStatus)[keyof typeof OperatorStatus];

export const OperatorStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Operator {
  id: number;
  name: string;
  description: string;
  type: OperatorType;
  status: OperatorStatus;
  version: string;
  createdAt: string;
}

export type GithubCallbackParams = {
  code: string;
  state?: string;
};

export type ListEvidenceParams = {
  /**
   * @nullable
   */
  teamId?: number | null;
  /**
   * @nullable
   */
  deploymentId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
};