# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Platform Engineering Control Plane API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /namespaces:
    get:
      operationId: listNamespaces
      tags: [namespaces]
      summary: List namespaces (admin sees all, user sees own team namespaces)
      responses:
        "200":
          description: List of namespaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Namespace"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createNamespace
      tags: [namespaces]
      summary: Create a namespace for a team and environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNamespaceBody"
      responses:
        "201":
          description: Namespace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Namespace"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /namespaces/{namespaceId}:
    get:
      operationId: getNamespace
      tags: [namespaces]
      summary: Get namespace details
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Namespace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Namespace"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNamespace
      tags: [namespaces]
      summary: Delete a namespace
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Namespace deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /starterkits/{starterKitId}:
    get:
      operationId: getStarterKit
      tags: [starterkits]
      summary: Get starter kit details
      parameters:
        - name: starterKitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Starter kit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StarterKit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments (admin sees all, user sees team deployments)
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Trigger a deployment (runs compliance checks first)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentBody"
      responses:
        "201":
          description: Deployment created (may be blocked by compliance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deployments/{deploymentId}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence records (admin sees all, user sees team evidence)
      parameters:
        - name: teamId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: deploymentId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of evidence records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenceRecord"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /evidence/{evidenceId}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get evidence record details
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Namespace:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        teamSlug:
          type: string
        environment:
          type: string
          enum: [dev, prod]
        k8sNamespace:
          type: string
        status:
          type: string
          enum: [active, provisioning, terminating]
        createdAt:
          type: string
          format: date-time
        resourceQuota:
          type: object
          properties:
            cpuLimit:
              type: string
            memoryLimit:
              type: string
            podLimit:
              type: integer
          required:
            - cpuLimit
            - memoryLimit
            - podLimit
      required:
        - id
        - teamId
        - teamName
        - teamSlug
        - environment
        - k8sNamespace
        - status
        - createdAt
        - resourceQuota
    CreateNamespaceBody:
      type: object
      properties:
        teamId:
          type: integer
        environment:
          type: string
          enum: [dev, prod]
      required:
        - teamId
        - environment
    StarterKit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [web, api, worker, ml, data]
        language:
          type: string
        framework:
          type: string
        tags:
          type: array
          items:
            type: string
        dockerfileIncluded:
          type: boolean
        cicdIncluded:
          type: boolean
        complianceReady:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - language
        - framework
        - tags
        - dockerfileIncluded
        - cicdIncluded
        - complianceReady
        - createdAt
    Deployment:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        namespaceId:
          type: integer
        namespaceName:
          type: string
        environment:
          type: string
          enum: [dev, prod]
        serviceName:
          type: string
        imageTag:
          type: string
        status:
          type: string
          enum: [pending, running, blocked, failed, succeeded]
        complianceStatus:
          type: string
          enum: [pending, passed, failed]
        testCoverage:
          type: ["number", "null"]
        triggeredBy:
          type: integer
        triggeredByUsername:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - teamName
        - namespaceId
        - namespaceName
        - environment
        - serviceName
        - imageTag
        - status
        - complianceStatus
        - testCoverage
        - triggeredBy
        - triggeredByUsername
        - createdAt
        - updatedAt
    CreateDeploymentBody:
      type: object
      properties:
        namespaceId:
          type: integer
        serviceName:
          type: string
        imageTag:
          type: string
        testCoverage:
          type: ["number", "null"]
      required:
        - namespaceId
        - serviceName
        - imageTag
    EvidenceRecord:
      type: object
      properties:
        id:
          type: integer
        deploymentId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        environment:
          type: string
          enum: [dev, prod]
        serviceName:
          type: string
        imageTag:
          type: string
        checkType:
          type: string
          enum: [test_coverage, security_scan, policy_gate, rbac_check]
        status:
          type: string
          enum: [passed, failed, skipped]
        details:
          type: string
        threshold:
          type: ["number", "null"]
        actual:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deploymentId
        - teamId
        - teamName
        - environment
        - serviceName
        - imageTag
        - checkType
        - status
        - details
        - threshold
        - actual
        - createdAt
...